requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
celery==5.2.7
redis==4.5.1
gunicorn==20.1.0

requests==2.28.2
bs4==0.0.1
lxml==4.9.2
fake_useragent==1.1.1
selenium==4.8.2
python_dotenv==1.0.0
